Clone BuildData before making changes
The BuildData from the previous build should be cloned before any
modifications are made.  If the build data from the previous build is
modified it will display incorrectly in the previous build until Hudson is
restarted and is able to load the correct BuildData object from the
build.xml file.