Add simple usecase support to BuildChooser
Allow a simple use case with less complexity if the user is only building
changes from a single branch in a single repository.  The simple usecase
includes two features that don't make sense for the advanced usecase:

* The chosen "branch" can be a commit ID.  This can be used if the user
  wants to build a particular commit that is not identified by a branch.

* The chosen "branch" can include build parameters.  Build parameters are
  not available during SCM polling, so this is only applicable to manually
  executed builds.

To support these changes, the automatic addition of "*/" to a branch
specifier had to be removed.  This was needed so that if a commit ID could
be entered without "*/" being prepended to the commit ID.

In addition wildcard support has been enhanced to an Ant path like syntax.
* matches any number of characters other than "/" and ** matches any number
of any characters.  This ensures that a specifier like "*/master" only
matches branches named master and not all branches that end in "/master".
Some advanced users may need to adjust their branch specifiers as a result
of this change.