There is a failure mode that causes the GIT plugin to be stalled.
If the plugin successfully clones a remote repository that has no initial
commit, then the value of HEAD is undefined (though the .git directory exists).

This can happen with things like gerrit repos that have content pushed to
them, but do not have a 'master' branch yet (HEAD maps to master, but
master does not exist).

When subsequent builds are attempted, the plugin thinks that the presence
of .git means the repo has been successfully cloned. However, git fetch will
fail, because there is no HEAD reference.

Alter the plugin to additionally see if the HEAD commit is valid
(git rev-parse --verify HEAD) and treat it as not valid if the return
generates an error, allowing a re-clone.

Signed-off-by: Nigel Magnay <nigel.magnay@gmail.com>