Reimplemented SCM option storage for automatic binding
Previously the communication between the html form and the java objects
was manually implemented by searching the request object for values by
full tag. This means that placing the GitSCM inside the MultiSCM plugin
would break binding of form values to the GitSCM object.

Also replaced storage of jgit configuration with the newly introduced
UserRemoteConfig and UserMergeOptions classes, since storage of
RemoteConfig objects would lead to an empty config when placed in
MultiSCM.

Note that this introduces the subtle difference that jgit options are
determined from user options at each GitSCM instantiation instead of
having them determined when the configuration is saved.

Current issues:

- merge options are not handled properly. Unfortunately the jelly format
  is not documented, so I have no idea on how to fix this now.
- some tests still fail, although the failures are non-deterministic.
  This points to some hidden inter-test dependency, or improper test
  setup/teardown.