Fixed matrix builds of git-based multi-scm configurations
The git plugin fetches the revisions to build from the parent job.
In a multi-scm environment, multiple build data objects are available in
the parent job. The second GitSCM would fail because it would attempt
to fetch a revision only present in the first GitSCM repository.

Added as proof-of-concept a new user-defined field discriminating the
SCMs in a multi-scm environment, making it possible for each GitSCM to
refer to its own build data.