A user on IRC was reporting a failure case where their server was UNIX and their client was Windows, which appeared to show the plugin bombing on a clone() operation.
Investigation showed that it was probably likely to the machinery that
tries to work out the version number thinking that 2 lines had been
returned, and generating an exception. This could be a CRLF / slave platform
difference, but I cannot recreate it.

Since the entire purpose is simply to work out if git clone supports --progress,
which is a nicety anyway, default to assuming it doesn't support it and
just log an error rather than crashing the entire build.

Signed-off-by: Nigel Magnay <nigel.magnay@gmail.com>