Add the ability to control whether submodules are updated
My repository can require special processing on the submodules before
they are updated. This is because Gerrit is used, so the submodules
may not always point to refs/heads/, which is the only place git looks
at by default. This should really be fixed in git, but as a workaround
we have created a script that checks ls-remotes and updates to the
appropriate location.

This shouldn't be used in normal circumstances, so I would consider
this somewhat of a hack; however, the option itself doesn't require any
obtrusive code in the git plugin.