Revert to using firstLine in validateRevision.
The revision hash should come out in the first line of git rev-parse.
This changes the firstLine method so that it will log when multiple
lines are returned, rather than failing. This works around the
"Process leaked file descriptors" problem that frequently occurs on
Windows. (Since this effectively causes a multi-line result.)