keep building when there are multiple candidate revisions
Multiple candidate revisions probably means that you missed a commit notification.
This can happen because Jenkins is down, or the notifier was flaky.

In this case, the git plugin currently builds the oldest, but we should also
schedule another build of the project so that we don't permanently stay "behind"
the latest.