Avoid some calls to "git show" in isRevExcluded
If there are no inclusion patterns, no exclusion patterns and no excluded
users, then the revision cannot be excluded.  If the revision cannot be
excluded, there is no reason to call "git show" to generate the list of
revisions to be tested for exclusions.

Provides some performance improvement in almost all cases which do
not define inclusion or exclusion patterns or users.  Will provide
significant performance improvement in those cases where the "git show"
output is very large (or hangs when being generated).