Retry on git exception
Usually github is not available for a few seconds. During the 24 hours period it could cause a lot's of issues. And it causes for us. Combining this with the scmCheckoutRetryCount we can avoid the failed builds caused by the github temporary outage. The original exception is only shown for the last iteration.