Moving out the clean part from the if/else branches.
This allows me to then extract this out into a separate extension point

This changes the order of commands for the merge case, which technically makes this change an incompatible change.

I still think this change is OK, because the point of the clean operation is to prevent the left-over from the previous build from affecting the next build. And certainly if the workspace is dirty, the merge operation can fail.