Disabled the JGit/CGit selection option.
I don't think it's desirable to have two places to let the user specify which git to use (one is this switch and the other is GitTool.)

I just made a change to the git-client plugin to allow JGit to be added as a Git installation as a first-class citizen, and I think that provides more logically consistent configuration. In case ndeloof objects to this change, I'm just commenting out the code as opposed to deleting it.