[FIXES JENKINS-19022] Remove branch-to-build map from build objects
This commit fixes the issues reported in JENKINS-19022. It removes the
need to store the branch-to-build map in the builds themselves. Instead,
the data is now stored centrally, thus avoiding the excessive overhead
of attaching an eternally growing history of branches to each build.

This can easily save several hundred megabytes of RAM and disk-space on
"grown" Jenkins/GIT deployments.