Fixed remaining test issues.
All 129 tests now succeed. The failing "slave" tests were due to the
fact that the SCM and BuildData objects are serialised and sent to the
remote machine for execution. Since the fields were static/transient,
the test failed as no data was transmitted.

Now, the BuildData class contains a serialisable reference to the static
object stored in the GitSCM class. It is not saved to disk via XStream,
but can be copied via the network to the build host. Since the structure
is only read, but not changed on the build host, that is acceptable.

Do note though, that the structure will still grow with each branch,
thus eventually slowing down the test execution.