[FEATURE] Introduce entry field for git merge flag
This patch introduces an entry field as part of the user merge options.
The field is meant to receive a flag that gets passed to the git merge
command (e.g. git merge --no-ff ... ). The extracted flag string is finally
stored in a enviroment variable GIT_MERGE_FLAG which needs to be picked
up from the git-client plugin (null and "" being ignored).