[FIXED JENKINS-14026] git plugin doesn't support branch name contains slash ("/")
The main issue is to determine whether an input containing a slash is a
qualified branch (remote followed by branch name) or a branch name on
which the remote name needs to be prepended.

When there is only one remote involved, that can be solved by prepending
the remote name to the branch name if the branch name doesn't already
start with the remote name.

Note : This commit doesn't cover cases where the given branch name is a
pattern