Enable use of trackingSubmodules
trackingSubmodules were introduced in v1.8.2 of GIT as a way to always
track the tip of the branch you are interested in for a submodule.  This
is turned on via the --remote switch to 'git submodule update'.  This
allows you to always stay up to date (--remote) or to just use the
specific SHA-1 you last downloaded (no --remote).

So, we now surface this new feature in the UI.  It requires corresponding
changes to the git-client-plugin (Pull request 122) for this to compile.