Document and improve test for building a sha1 rev
The 'branch to build' option already also accepts a SHA1-hash to specify
the specific revision to build, but this was not really obvious (neither
in the online help nor in the code). Added a note to the documentation and
a test which covers part of this functionality more explicitly, though part
of this still happens in GitSCM which isn't really under test.

See also JENKINS-22278