Fix to compareRemoteRevisionWithImpl(...)
String singleBranch not used when checking for head revision.
If the Git branch name is represented by an environment variable the variable name is passed to git.getHeadRev(...), instead of the expanded value in singleBranch.

This causes head to always be null, and the build is triggered continuously after polling, without any changes being present in the branch.