Revert "Depend on git-client 1.10.2" until test failure fixed
This reverts commit 87beb7d0270c3058c90fd178b0aa96049ad6a900 until
the automated test which fails in the code can be fixed.  The test is
specifically noted that it is an area of bugs in the prior client plugin,
and the fix of JENKINS-23299 in git client plugin 1.10.2 has improved
the behavior, but the test needs to be updated to show the new behavior
(and have its TODO comment removed).