Revert "Update git-client dependency to 1.10.2"
This reverts prior commit until the automated test which fails in the
code can be fixed.  The test is specifically noted that it is an area
of bugs in the prior client plugin.  The fix of JENKINS-23299 in git
client plugin 1.10.2 has improved the behavior, but the test needs to
be updated to show the new behavior (and have its TODO comment removed).