Check for matching branches in specific remotes first
When a job is configured with two git remotes, always check that the
branch specifier matches the given remote (if applicable).  If no remote
matches, revert to the previous behavior of trying the branchSpec in
every remote.

This fixes a problem where:
  1. Remote origin contains a branch upstream/master
  2. Remote upstream contains a branch master
  3. User specifies branch: "remotes/upstream/master"

Before this commit, the job would build "remotes/origin/upstream/master", even
though the user said to build "remotes/upstream/master".