Actually skip tests for jgit poller
Tests are already ignored but appear as if they pass.

Specifically running with the junit 4 runner and using the @ignore annotation to actually skip the tests

Added an override for all inherited junit 3 tests and delegated.

Added an additional test to fail if a test gets added and not overridden as this would result ina silent ignore