add Junit4 adapter to contain boiler-plate
The adapter overrides all junit3 tests and adds the @Test annotation and deligates. This is purely boiler-plate as it also calls the inherited setup and teardown methods.

http://stackoverflow.com/questions/1635108/adding-java-annotations-at-runtime

Converted two junit 3 tests to use the new adapter and now 'nicely' skip the tests.

If new junit3 tests are added to GitSCMTest then they would be silently skipped by the junit4 runner because they are not annotated. To address this i've added a test to check sctively look for inherited junit3 tests and confirm they are overridden by the adapter.