[JENKINS-27315] Add environment variable for git commit author email
Surprised no one had added this before. If someone could show me an alternative
way  to get the commit's author's email, that would be great.
This checks the commit's text for the author line and parses out the email
address.
The environment variable is set as GIT_COMMIT_AUTHOR_EMAIL.
My team use this to email the person who comitted a piece of code at the end of
a  build.