Check BuildData.getLastBuild and BuildData.hasBeenBuilt
Improved assertion failure messages.

Initially used this test to explore the behavior of the getLastBuild
fast check.  Concluded that while the fast check was never succeeding,
it was also not causing incorrect results to be returned by the
getLastBuild(sha1) method.

Intentionally left the few extra assertions in the code rather than
delete them, since they are low cost to execute and might detect a
problem in the future.