[Fix JENKINS-28792] GitLab browser edit panel allow more values for version
When saving a job defined to use the GitLab browser, if the version field
is empty or contains a value which cannot be converted to a double,
the save operation fails with a stack trace that the Double.valueOf()
threw a NumberFormatException.

Silently use a default value if the user provided value cannot be
converted to a Java double.