This fixes a bug where the filtering removes all tipCandidates
This bug occurs when there are multiple matching tags on the same commit.
The existing logic removed the tipCandidate if the commit was already visited.
If there are two revisions on the same commit, the following happens:
The first run places itself in the tipCandidates and marks the commit as visited.
The second run (via the second revision) then places the same commit with another
revision in the tipCandidates but removes it right after because the commit is
already marked as visited.

This patch skips this misbehavior and just keeps the the first revision of a commit.