Remove the test for purge deleted branches from BuildData
Test did not fail on my primary configuration, but fails on
other configurations and fails on the CloudBees hosted job.
Since the code to support the test was already removed in
be1890377a4623be9483e6b70df22cd8da2f7616, retaining a test which depends
on removed code is not helpful.