[FIXED JENKINS-09713] Consider tags when checking which branches to build
When using multiple branch specifiers, the git plugin uses an advanced
selection mechanism as defined in `getAdvancedCandidateRevisons` which
does not consider tags from the repository, only normal branches.

E.g. when using the following Branch specifiers:

- refs/remotes/origin/${B}
- refs/tags/${B}

where B is a build parameter having a default value of "master" never
causes any tag to be build, regardless of what the user chooses.

Include the tags to complete the list of valid branches to find the best
matching ref for the current build.

Add testMultipleBranchesWithTags to GitSCMTest

This test checks that for multiple branch specifications including some
tag references, the build succeeds and selects the correct revision.

Pull request #340