[Fix JENKINS-30073] changeset time shouldn't be -1
The changeset time parsing implemented in an earlier commit did not
account for the most common format of the "nearly ISO" date format
generated by the "+%ci" format argument.

The fully ISO 8601 compliant date format argument ( +%cI ) is not
available in all the git versions supported by the plugin, so the plugin
continues to use the nearly ISO 8601 compliant format ( +%ci ), then
transforms it into an ISO 8601 format when the timestamp is requested.

For example, git 2.1.4 as shipped with Debian 8 does not recognize +%cI
for date formatting.  Git 2.6.0 version does recognizes +%cI.