Insert git 1.7.1 SHA1 exclusions list into test
The plugin officially requires git 1.7.10 or later for credentials
support, but CentOS 5, CentOS 6, and some other distributions deliver
git versions older than 1.7.10.

This specific problem is that the "git whatchanged" command on git
1.7.1 on CentOS 6 fails to report any output for a few of the commits
in the git-plugin repository.  The "git whatchanged" command doesn't
seem to have that behavior on any of the other platforms which run
these tests, only on CentOS 6.7.

Rather than keep the test simple and exclude execution on CentOS 6.7,
I created a list of the SHA1's to be skipped so that all the other
SHA1's can be tested on CentOS 6.7.  The exclusion also only applies
to command line git, not to JGit, so those SHA1's will be tested on
CentOS 6.7 in the JGit implementation.