git-plugin: add support to ignore errors from GitPublisher
Primarily to allow the usecase of ignoring the failure when pushing to a
gerrit refspec without pushing new changes, add an option
"ignoreGitErrors" to the GitPublisher which will enable creation of a
publisher that will simply keep going when receiving errors from git
commands. Other errors not the result of a GitException will still be
handled. In addition, the stack trace will still appear on the logs but
won't fail the builds.

Add a test for the same based on force push, that ensures the build
succeeds but does not actually result in a completed push.