JENKINS-24728 Added UI Elements for shallow clone depth
Use @DataBoundSetter instead of extending the constructor as was done
in the original pull request.  The DataBoundSetter annotation simplifies
the change.

Original concept by Nicholas Jasieniecki, with changes by Mark Waite.

Add help to shallow clone depth setting.

Add shallow and shallow depth to fetch command options.  The previous
implementation only set depth on the clone arguments, not on the fetch
arguments.  Without the depth on the fetch arguments, fetch retrieves
history additively, rather than retaining the depth at the originally
specified value.  That caused the history in a repository which is being
reused to grow each time new commits are detected and built by the job.
This change now keeps the history at the specified depth even with
incremental updates of the repository by new commits.