Test JENKINS-30178 using GitStatus.toString
Asserts that job parameter default values are available when a job
is started by a notifyCommit.  If the notifyCommit includes a sha1
parameter, then the changes for JENKINS-27092 fail to assign parameters
their default values.

Modifying the GitStatus object to be more easily tested was simpler
than using a TestExtension.  Should eventually replace the testing
misuse of the GitStatus.toString() method with a TestExtension of
GitStatus.Listener.