git: throw IOException around submodule update failures
Since these commands can fail due to the same reason as git checkout,
(too many concurrent accesses to the git server, for example), we need
to throw an exception caught by the retry loop. Based on inspection, it
appears that the best choice to catch is GitException, since that's what
is thrown by the launchCommand flow that calls the git submodule
command.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>