[JENKINS-31934] throw IOException around submodule update failures
Updating the submodule can fail in a similar manner to git checkout of
the parent project. If we simply throw the GitException, then the
generic retry loop will not kick in.

For now, the only exceptions we'll catch is the GitException which
should be generated if the launchCommand fails.

Without this change, intermittent failures when attempting to pull
submodule data from a git remote will not properly allow the retry loop
to kick in.

Add a new test case to verify that we actually do throw the IOException.
Because this exception case is likely due to an intermittent error,
we'll use mocks to setup a case where we always throw the exception.

Signed-off-by: Jacob Keller <jacob.e.keller@intel.com>