[SECURITY-275] Honor parameters declared safe
If the Jenkins administrator wants to allow certain parameters to be
passed from the notifyCommit GET request to their jobs, they define those
parameter names as the comma-separated string value of the property
hudson.plugins.git.GitStatus.safeParameters or as the comma-separated
value of the property hudson.model.ParametersAction.safeParameters.

The default is to pass none of the parameters from the GET request to
the jobs.