[JENKINS-36507] Move refspec control of initial clone to CloneOption
The decision if the refspec should be honored during clone is now
part of the CloneOption additional behaviour.  The user can control
the setting per job, and the default setting remains compatible with
the git plugin 2.5.0 (and prior) cloning of all references initially,
then later honoring the refspec in other operations.

Prior to git plugin 2.5.1, JENKINS-31393 caused the user provided
refspec to be ignored during the initial clone. It was honored in
later fetch operations, but not in the first clone. That meant the
initial clone had to fetch all the branches and their references from
the remote repository, even if those branches were later ignored due
to the refspec.

The fix for JENKINS-31393 exposed JENKINS-36507 which indicates that
the Gerrit Plugin assumes all references are fetched, even though it
only passes the refspec for one branch. The git plugin shouldn't
require a code change in the gerrit plugin, so each job must choose if
JENKINS-31393 will affect that job or not.

Most jobs that are significantly affected by JENKINS-31393 will likely
need to make other changes in the "Advanced clone options" (like using
a reference repository, or disabling the cloning of tags, or enabling
shallow clone).

Added CloneOption tests for JENKINS-36507 using the same basic test as
was earlier submitted for JENKINS-31393.

The testSpecificRefspecs and testSpecificRefspecsWithoutCloneOption
methods confirm behaviors of refspecs on initial clone. Without the
CloneOption to honor refspec, all references are cloned, even if they
will be later ignored due to the refspec.  With the CloneOption to ignore
refspec, the initial clone also honors the refspec and only retrieves
references per the refspec.