Use reflection to find the close method
Java 8 (and possibly others) refuses to runtime cast a RevWalk to a
Closeable, even when the implementation loaded into the current class
loader is a Closeable.  I assume the cast is resolved at compile time
as possibly acceptable, but then runtime type safety checks reject it
before evaluating the currently loaded RevWalk implementation.

The verbose logging is temporary while I confirm with runtime testing
that the various scenarios are handled correctly.

Scenario tested (and working as far as I can tell):

* JDK 8 linux, git client 2, git plugin built with git client 1.19

Scenario to be tested:

* JDK 8 windows, git client 2, git plugin built with git client 1.19
* JDK 7 linux, git client 2, git plugin built with git client 1.19
* JDK 7 windows, git client 2, git plugin built with git client 1.19