Use credentials 2.1.7 - don't force users to 2.1.8
The Jenkins credentials plugin wiki page currently notes that credentials
plugin 2.1.8 added some additional diagnostics and safety which may expose
bugs in other plugins, and may require a restart of the Jenkins server.

If the git plugin depends on 2.1.8, then the user is required to update
to 2.1.8.  Depending on credentials plugin 2.1.7 still allows the user
the option of running 2.1.8, but will also allow them to remain on 2.1.7.