Normalize BuildData URL comparisons to reduce duplication
A trailing slash on a Git repository URL does not generally change
the URL, nor (in most cases) does a trailing ".git" suffix.  BuildData
comparisons are now done by removing any trailing slashes, and removing
the .git suffix (if it exists), and then (if a URL form like http://
or https:// or ssh://) normalizing the URL with URI.normalize().

The normalization does not attempt to reconcile different protocols which
point to the same repository, like https://github.com/jenkinsci/git-plugin
and git@github.com:jenkinsci/git-plugin.