Initialize global git defaults in tests, if unset
Explicitly set command line git global config values

When a job needs to tag a build, command line git requires a setting for
user.email and user.name.  There doesn't seem to be an easy location to
insert that into the repository clone process without risking greater
disruption to the code structure, so this sets the global configuration
at the start of the test.

Global configuration values are only set for those cases where no value
is already assigned.  If a value is already assigned, then it is left
unchanged.