Option for resolving matrix commit IDs when they run
By default, matrix commit IDs are resolved on the master run, before each matrix run is executed. This has a major disadvantage for pull requests where the commit hash can become invalid if there are new commits between the matrix run and when the commit ID was resolved at the master node. Enabling this option can introduce build inconsistency across build configurations, but avoids the issue where checkout fails for pull requests if there are commits between the matrix run and when the master node is executed.