Broaden retrievePrimaryHead tests for git versions
Git versions prior to 2.8.0 do not include the "--symref" argument to the
"git ls-remote" command.  In that case, the code uses a heuristic that
if the SHA1 of the remote HEAD matches exactly the SHA1 of one of the
remote branches, that is the default remote branch.

The heuristic breaks down if more than one remote branch point to the
same SHA1 as the remote HEAD.  Git versions 2.8.0 and later correctly
report the remote default branch, while earlier versions (and JGit)
will repo no default.

Unfortunately, most major Linux distributions do not yet distribute Git
2.8.0 or later.  That includes the current Jenkins docker image, which
is based on Debian 8.