Configure global git user email & name if not configured
Newly created machines and accounts may not yet have a setting for the
git user.name and user.email properties.  That causes GitPublisherTest,
GitSCMTest, and GitStepTest to fail.

This change will check if a global value is already configured for
user.name and for user.email.  If either is not configured, it will
configure a value.  If a value is already .configured, it is left
unmodified.