JENKINS-44640 - Fix NPE checking legacy git SCM URL
Credential tracking is intended to track relevant and valuable use of
the credential, not something as simple as using that credential in a
form validation.

Form validation is valuable for the user experience of the
administrator, but it is not valuable enough to track the credential
use in validating the form.