Allow disabling of multi candidate scheduling
Add an additional behaviour that disables the multi candidate
scheduling.

The default behaviour of the git plugin is to check for new changes
and trigger a new build for every tip of every branch.
When working with big repositories, high build times and many
independent branches, this behaviour leads to an unnecessary high
resource consumption.