[JENKINS-45729] Move printCommitMessageToLog after checkout, ignore exceptions
JENKINS-45729 shows that there are cases where the expected commit is not
yet available to be printed or some other issue is causing an exception.

Printing the commit message of the last commit to the log file is a
diagnostic aid, not critical to success of the build.  An exception in
a diagnostic should not break the build.

I'm unable to duplicate JENKINS-45729 in any consistent way.  I have
seen the reported exception at least once, but the project with that
exception stopped showing that exception.