[JENKINS-9016] Added an option to search for users based on email attribute
This option is added in order to try to reduce unnecessary user
creation when user ids in jenkins are not based on email addresses.
It will allow optional search for users based on their mail address
property, assuming a single mail address should match only one user.

This is sort of a common sense implementation allowing to wait for
the implementation of JENKINS-14849, as the actual target, requiring
SecurityRealm plugins to expose the logic to match SCM ids to the
users belonging to the realm.