Use Jenkins 2.60.3 and Java 8 as base version
Update to use pre-release of git client plugin 3.0.0 that uses Jenkins
2.60.3 and Java 8 as base version.

Also test with Jenkins 2.107.2 for compatibility check.

Intentionally choosing Jenkins 2.60.3 since it is the first LTS to require
Java 8.  Allow the greatest number of installations to upgrade to the new
git plugin version while still allowing Java 8 capabilities in the plugin.