Add tests to show arbitrary nature of requiresWorkspaceforPolling
If the branch name is parameterized and the parameter is empty,
then the workspace is not required for polling.

If the branch name is empty, then a workspace is required.

If the branch name contains a wildcard that is not '*/' in the
first two characters, a workspace is required.

If multiple branch names are provided, a workspace is required.

If branch name is parameterized and parameter includes a wildcard,
a workspace is not required.